src/pyproject.toml
[project]
name = "parlor"
version = "0.1.0"
description = "On-device, real-time multimodal AI — have natural voice + vision conversations, fully local and private"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.135.3",
    # TTS: kokoro-onnx runs on Linux + Windows; macOS uses mlx-audio instead.
    "kokoro-onnx>=0.5.0; sys_platform != 'darwin'",
    # LLM: litert-lm has no Windows wheels (only manylinux + macOS arm64).
    # Windows uses the transformers-based backend in llm.py instead.
    "litert-lm>=0.10.1; sys_platform != 'win32'",
    "misaki[en]>=0.9.4; sys_platform == 'darwin'",
    "mlx-audio>=0.4.2; sys_platform == 'darwin'",
    "num2words>=0.5.14; sys_platform == 'darwin'",
    "numpy>=2.4.4",
    "pillow>=11.0.0",
    "soundfile>=0.13.1",
    "uvicorn[standard]>=0.43.0",
    "websockets>=16.0",
    # Windows-only LLM backend: HuggingFace transformers for Gemma multimodal,
    # faster-whisper for speech-to-text. These ship Windows wheels.
    "torch>=2.4.0; sys_platform == 'win32'",
    "transformers>=4.48.0; sys_platform == 'win32'",
    "accelerate>=1.3.0; sys_platform == 'win32'",
    "faster-whisper>=1.1.0; sys_platform == 'win32'",
]

# Ensure uv resolves wheels for every supported platform, including Windows.
# Without this, `uv sync` on Windows fails on packages whose wheels exist for
# other platforms only (e.g. litert-lm-api).
[tool.uv]
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
]
